refactor(routes): move auth domain into routes/auth/ subpackage
Slice 2n of the route-domain reorganization (#4082/#4071, per
specs/architecture-runtime-inventory.md §6.3). Moves auth_routes.py
(836 lines), api_token_routes.py (209 lines), and device_flow.py (193 lines)
into routes/auth/, leaving backward-compat sys.modules shims at the old
paths. Pure file reorganization, no behavior change.

All three shims use sys.modules replacement so:
- sys.modules.pop("routes.auth_routes") + re-import (test_auth_policy.py)
- monkeypatch.delitem(sys.modules, "routes.api_token_routes") + re-import
  (test_api_token_routes.py)
- import ... as ar + monkeypatch.setattr(ar, "DEEP_RESEARCH_DIR", ...)
  (test_rename_user_owner_sync.py)
- monkeypatch.setattr(device_flow, "require_admin", ...)
  (test_device_flow_routes.py)
all reach the canonical modules.

The 3 auth files do NOT import each other — they are independent modules
that share the auth domain grouping. device_flow is also imported by
copilot_routes and chatgpt_subscription_routes (external); the shim keeps
those working unchanged.

Zero source-introspection landmines. SESSION_COOKIE constant imported by
app.py resolves automatically through the sys.modules shim.

Adds tests/test_auth_routes_shim.py to pin all 3 sys.modules shim contracts.

Verified: compileall clean; full suite 4806 passed, 3 skipped.
