+++ routes/auth/__init__.py
@@ -1,0 +1,6 @@
+"""Auth route domain package (slice 2n, #4082/#4071).
+
+Contains auth_routes.py, api_token_routes.py, and device_flow.py, migrated
+from the flat routes/ directory. Backward-compat shims at the old paths
+re-export from here.
+"""
